+++ src/Header.jsx
@@ -1,5 +1,4 @@
 import React from "react";
-import clsx from 'clsx';
 import { makeStyles } from '@material-ui/core/styles';
 
 import {
