+++ src/UseStyles.jsx
@@ -1,0 +1,21 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles(function (theme) {
+	return {
+		container: {
+			paddingTop: theme.spacing(4),
+			paddingBottom: theme.spacing(4),
+		},
+		paper: {
+			padding: theme.spacing(2),
+			display: 'flex',
+			overflow: 'auto',
+			flexDirection: 'column'
+		},
+		fixedHeight: {
+			height: 240
+		}
+	};
+});
+
+export default useStyles;
